Add a scheduled task to send email reminders for expiring passwords
Introduced functionality to notify users about expiring passwords via email. Added a scheduled task to automate this process, along with the corresponding configurations and repository methods.

+++ src/main/java/fi/asteriski/nakitin/service/EmailService.java
@@ -56,7 +56,7 @@ public class EmailService {
         }
     }
 
-    public void sendPasswordExpirationWarning(String toEmail, int daysUntilExpiration) {
+    public void sendPasswordExpirationWarning(String toEmail, long daysUntilExpiration) {
         var subject =
                 messageSource.getMessage("email.subject.password-expiration", null, LocaleContextHolder.getLocale());
         var message = messageSource.getMessage(
